requirements.txt
Flask
mysql-connector-python
flask-cors
sentence-transformers
